% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_iter}
\alias{batch_iter}
\title{make batch datasets, each with size batch_size}
\usage{
batch_iter(data, batch_size)
}
\arguments{
\item{data}{A dataset}

\item{batch_size}{Batch size}
}
\value{
This function returns \code{num_batch} datasets with size equals \code{batch_size} each
}
\description{
make batch datasets, each with size batch_size
}
