% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNA.R
\name{createNA}
\alias{createNA}
\title{Create missing value for a dataset}
\usage{
createNA(data, p = 0.1, seed = 2019)
}
\arguments{
\item{data}{a dataframe}

\item{p}{the proportion of missing values in the dataframe}

\item{seed}{random seed, a singlue value}
}
\description{
Create missing value for a dataset
}
