% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midae-initialization.R
\name{denoise_decoder}
\alias{denoise_decoder}
\title{Midae Decoder update function}
\usage{
denoise_decoder(act, z, weights, biases, hidden_drop, decoder_structure)
}
\description{
Midae Decoder update function
}
