% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midae-initialization.R
\name{denoise_encoder}
\alias{denoise_encoder}
\title{Midae Encoder update function}
\usage{
denoise_encoder(
  act,
  x,
  weights,
  biases,
  input_drop,
  hidden_drop,
  encoder_structure
)
}
\description{
Midae Encoder update function
}
