% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Midaeboot-imputer.R
\docType{class}
\name{Midae.boot}
\alias{Midae.boot}
\title{R6 class for Midae bootstrapping imputer}
\description{
Set up a Midae (with bootstrapping) imputer object with specified hyperparameters and then obtain multiple imputed datasets
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Midae.boot$new()}}
\item \href{#method-train}{\code{Midae.boot$train()}}
\item \href{#method-impute}{\code{Midae.boot$impute()}}
\item \href{#method-clone}{\code{Midae.boot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Midae.boot$new(
  data,
  n_h = 2L,
  encoder_structure = c(128L, 64L, 32L),
  decoder_structure = "reversed",
  learn_rate = 1e-04,
  iteration = 400,
  batch_size = 500,
  input_drop = 0.2,
  hidden_drop = 0,
  act = tf$nn$elu,
  weight_decay = "default",
  epsilon = 1e-07,
  rho = 0.95,
  decay = 0.9,
  momentum = 0.9,
  optimizer = "Adam"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Midae.boot$train(verbose = TRUE, print_freq = 1, m = 5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-impute"></a>}}
\if{latex}{\out{\hypertarget{method-impute}{}}}
\subsection{Method \code{impute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Midae.boot$impute(
  m = 5,
  onehot = FALSE,
  all.numeric = FALSE,
  add.noise = FALSE,
  SD = 1,
  pmm = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Midae.boot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
