% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Midae-imputer.R
\docType{class}
\name{Midae}
\alias{Midae}
\title{Midae imputer object
R6 class for Midae imputer}
\format{An [R6Class] midae imputer object}
\usage{
Midae
}
\value{
[Midae]
}
\description{
Midae imputer object
R6 class for Midae imputer
}
\keyword{datasets}
