% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midae-initialization.R
\name{midae_output}
\alias{midae_output}
\title{Midae output evaluation}
\usage{
midae_output(act, x, network_weights, encoder_structure, decoder_structure,
  input_keep, hidden_keep)
}
\description{
Midae output evaluation
}
