% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scaler.R
\name{minmax_scaler}
\alias{minmax_scaler}
\title{scale a dataset and return a scaled dataframe, the colmin and colmax of each column}
\usage{
minmax_scaler(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
scale a dataset and return a scaled dataframe, the colmin and colmax of each column
}
