% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-misle.R
\docType{package}
\name{misle}
\alias{misle}
\alias{misle-package}
\title{mivae: multiple imputation using variational autoencoders}
\description{
scalable multiple imputation
}
\author{
\strong{Maintainer}: Yongshi Deng \email{yongshi.deng@auckland.ac.nz}

Other contributors:
\itemize{
  \item Thomas Lumley \email{t.lumley@auckland.ac.nz} [thesis advisor]
}

}
