% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mivae-imputer.R
\docType{class}
\name{Mivae}
\alias{Mivae}
\title{Mivae imputer object
R6 class for Mivae imputer}
\format{An [R6Class] mivae imputer object}
\usage{
Mivae
}
\value{
[Mivae]
}
\description{
Mivae imputer object
R6 class for Mivae imputer
}
\keyword{datasets}
