% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{mivae_optimizer}
\alias{mivae_optimizer}
\title{mivae cost function}
\usage{
mivae_optimizer(x, na_idx, networkOutput, output_split, output_struc,
  kld = TRUE, loss_weight = 1, kld_weight = 1, sigma_weight = 1)
}
\description{
mivae cost function
}
