% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mivae-initialization.R
\name{mivae_output}
\alias{mivae_output}
\title{Mivae output evaluation}
\usage{
mivae_output(
  act,
  x,
  network_weights,
  n_h,
  encoder_structure,
  decoder_structure
)
}
\description{
Mivae output evaluation
}
