% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xgb-imputer.R
\docType{class}
\name{Mixgb}
\alias{Mixgb}
\title{Mixgb imputer object
R6 class for Mixgb imputer}
\format{An [R6Class] mivae imputer object}
\usage{
Mixgb
}
\value{
[Mixgb]
}
\description{
Mixgb imputer object
R6 class for Mixgb imputer
}
\keyword{datasets}
