% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocess.R
\name{onehot}
\alias{onehot}
\title{This function is used to convert dataframe into onehot dataframe
It also returns \code{column_list}, each list component contains classes of one non-numeric feature}
\usage{
onehot(data, autolabel = TRUE, binary = NULL, multiclass = NULL)
}
\description{
This function is used to convert dataframe into onehot dataframe
It also returns \code{column_list}, each list component contains classes of one non-numeric feature
}
