% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocess.R
\name{output_structure}
\alias{output_structure}
\title{This function is use to identify the structure of input}
\usage{
output_structure(data, columns_list = NULL, binary = NULL,
  multiclass = NULL, unsorted = TRUE, verbose = FALSE)
}
\arguments{
\item{data}{A data frame (not one hot yet)}

\item{columns_list}{A list, each list component contains classes of one non-numeric feature}

\item{binary}{A vector contains names of binary features}

\item{multiclass}{A vector contains names of multiclass features}

\item{verbose}{A logical value indicating whether information should be printed out}

\item{unstorted}{A logical value indicating whether the data is sorted or not}
}
\value{
This function return a list with \code{size_index} and \code{output_struc}
}
\description{
This function is use to identify the structure of input
}
