% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocess.R
\name{sort_feature}
\alias{sort_feature}
\title{This function is used to sequence the columns of the dataset, so as to be in the order Numeric>Binary>Multiclass
It is used to minimise memory overhead}
\usage{
sort_feature(data, feature)
}
\arguments{
\item{data}{A data frame}

\item{feature}{name of the feature (character)}
}
\value{
This function returns a list with \code{sorted.df} and \code{chunk} (the number of feature in the selected group)
}
\description{
This function is used to sequence the columns of the dataset, so as to be in the order Numeric>Binary>Multiclass
It is used to minimise memory overhead
}
