% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mivae-initialization.R
\name{xavier_init}
\alias{xavier_init}
\title{Xavier Initialization using Uniform distribution}
\usage{
xavier_init(n_inputs, n_outputs, constant = 1)
}
\description{
Xavier Initialization using Uniform distribution
}
