% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mivae-imputer.R
\docType{class}
\name{Mivae}
\alias{Mivae}
\title{Mivae imputer object
R6 class for Mivae imputer}
\format{
An [R6Class] mivae imputer object
}
\value{
[Mivae]
}
\description{
Mivae imputer object
R6 class for Mivae imputer

Mivae imputer object
R6 class for Mivae imputer
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Mivae$new()}}
\item \href{#method-train}{\code{Mivae$train()}}
\item \href{#method-impute}{\code{Mivae$impute()}}
\item \href{#method-clone}{\code{Mivae$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mivae$new(
  data,
  n_h = 2L,
  encoder_structure = c(128L, 64L, 32L),
  decoder_structure = "reversed",
  learn_rate = 1e-04,
  iteration = 400,
  batch_size = 500,
  act = tf$nn$softplus,
  weight_decay = "default",
  epsilon = 1e-07,
  rho = 0.95,
  decay = 0.9,
  momentum = 0.9,
  optimizer = "Adam",
  kld_weight = 0.1,
  loss_weight = 0.5,
  sigma_weight = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mivae$train(print_freq = 1, bootstrap = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-impute"></a>}}
\if{latex}{\out{\hypertarget{method-impute}{}}}
\subsection{Method \code{impute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mivae$impute(
  m = 5,
  onehot = FALSE,
  all.numeric = FALSE,
  add.noise = FALSE,
  SD = 1,
  pmm = FALSE,
  pmm.k = 3
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mivae$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
